JavaVersion 11; use unittest-helper-0.5.1

+++ src/test/java/com/kazurayam/materialstore/util/CopyDirTest.java
@@ -24,8 +24,10 @@ public class CopyDirTest {
     @Test
     public void test_copyDirectory_overwriting() throws IOException {
         Path sourceDir = FixtureDirectory.getFixturesDirectory().resolve("issue#331");
-        Path targetDir = too.getClassOutputDirectory();
-        DeleteDir.deleteDirectoryRecursively(targetDir);
+        Path targetDir = too.resolveClassOutputDirectory();
+        if (Files.exists(targetDir)) {
+            DeleteDir.deleteDirectoryRecursively(targetDir);
+        }
         Files.walkFileTree(sourceDir, new CopyDir(sourceDir, targetDir));
         assertTrue(Files.exists(targetDir));
         Files.walkFileTree(sourceDir, new CopyDir(sourceDir, targetDir));
